Add basic notification functionality
When a user is tagged in an item, a notification is generated and the
count next to "Notifications" in the menu adds one. Users can read their
notifications by going to the notifications index. When a notification
is read, its count goes down. Implicitly, it will be deleted within one
week unless unread again. Styling, including icons, still needs to be
done.